Milo&KOS Switch Functionality and Condition Implemented
The switch functionality and condition have now been implemented, very
basic collectables have been added as well and is being randomly spawned
within the edges of the map. OBS! The Code might be target for
Refactoring in the future since I added a GameController Object which
should be responsible for the Game State.
Anyway Cheers! :D