Combined Controls of Sprint 01
Created a Scene containing the combined Controls created in Sprint 01.
Only the Jump animations seems to cause troubles when applied to the
existing Animator Controller, preventing movement and cause errors :(
It might be a Unity bug or something entirely different, but I can't
seem to fix the problem as of now.
The rest should work fine though, but you are more than welcome to have
a look at it yourselves :)
Cheers!